added auth to endpoints
